pdf-parser-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
groq==0.4.2
python-dotenv==1.0.0
pydantic==2.5.0 